HANUChatbot_BE/requirements.txt
psycopg2-binary~=2.9.9
flask~=2.2.5
flask-cors~=4.0.0
pandas~=2.1.4
pgvector~=0.3.2
python-dotenv~=0.21.0
openai~=1.9.0
tiktoken~=0.6.0